fw-4255, move predicates to new permissions package
