+++ firstvoices/backend/models/category.py
@@ -3,9 +3,9 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext as _
 
-from backend import predicates
 from backend.models.base import BaseSiteContentModel
 from backend.models.constants import CATEGORY_POS_MAX_TITLE_LENGTH
+from backend.permissions import predicates
 
 
 class Category(BaseSiteContentModel):
