+++ firstvoices/backend/models/part_of_speech.py
@@ -3,8 +3,8 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext as _
 
-from backend import predicates
-from backend.managers.permissions import PermissionsManager
+from backend.permissions import predicates
+from backend.permissions.managers import PermissionsManager
 
 from .base import BaseModel
 from .constants import CATEGORY_POS_MAX_TITLE_LENGTH
