+++ firstvoices/backend/models/sites.py
@@ -4,7 +4,7 @@ from django.core.validators import validate_slug
 from django.db import models
 from django.utils.translation import gettext as _
 
-from backend import predicates
+from backend.permissions import predicates
 
 from .base import BaseModel, BaseSiteContentModel
 from .constants import Role, Visibility
